Day 10 | Completed Part 2 - Changed Adapter Rating Type -> Long

+++ implementation/src/main/kotlin/io/github/tomplum/aoc/aircraft/adapter/Adapter.kt
@@ -2,8 +2,8 @@ package io.github.tomplum.aoc.aircraft.adapter
 
 import kotlin.math.abs
 
-data class Adapter(val rating: Int) {
-    fun ratingDifference(other: Adapter): Int = other.rating - rating
+data class Adapter(val rating: Long) {
+    fun ratingDifference(other: Adapter): Long = other.rating - rating
 
     fun isCompatible(other: Adapter): Boolean = abs(other.rating - rating) <= 3
 }
